+++ src/lib/components/notes/backlinks/backlinks.svelte
@@ -8,10 +8,12 @@
 <div class="w-full">
 	<h2 class="text-lg">Backlinks</h2>
 	<ul class="space-y-4 rounded-lg border border-gray-400 py-2 px-4">
-		{#each backlinks as link}
-			<li>
-				<BacklinkItem backlink={link} />
-			</li>
-		{/each}
+		{#if backlinks?.length > 0}
+			{#each backlinks as link}
+				<li>
+					<BacklinkItem backlink={link} />
+				</li>
+			{/each}
+		{/if}
 	</ul>
 </div>
