+++ src/routes/notes/tags/[tags].svelte
@@ -1,0 +1,8 @@
+<script context="module">
+	export const layout = false;
+</script>
+
+<script lang="ts">
+	export let success: boolean;
+	console.log(success);
+</script>
