+++ src/routes/notes/tags/[tags].ts
@@ -1,0 +1,10 @@
+import type { RequestHandler } from '@sveltejs/kit';
+
+export const get: RequestHandler = async ({ params }) => {
+	return {
+		status: 200,
+		body: {
+			success: true
+		}
+	};
+};
